travel_agency/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
